commit message

+++ source/mile_to_kilometer.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-double mile_to_kilometer(double mile)
+float mile_to_kilometer(float mile)
 {
 
     return mile * 1.609344;
@@ -8,7 +8,7 @@ double mile_to_kilometer(double mile)
 
 int main()
 {
-    double miles;
+    float miles;
     std::cout << "Enter mile: ";
     std::cin >> miles;
     std::cout << miles << " miles = " << mile_to_kilometer(miles) << "\n";
